feat: implement createPlatformVideo function and remove createTutorial function
